+++ app/src/main/java/com/example/pandas/ui/view/viewpager/Indicator.kt
@@ -25,8 +25,8 @@ public class Indicator : View, Iindicator {
     private var isMid: Boolean = false
 
     //距离参数，选中和未选中一样大小
-    private val indicatorItemPadding: Int = ScreenUtil.dip2px(4f) //间隔
-    private val indicatorRadius: Int = ScreenUtil.dip2px(2.7f) //半径
+    private val indicatorItemPadding: Int = ScreenUtil.dip2px(4f).toInt() //间隔
+    private val indicatorRadius: Int = ScreenUtil.dip2px(2.7f).toInt() //半径
 
     private var mPaint: Paint = Paint(Paint.ANTI_ALIAS_FLAG)
     private var selectColor: Int? = null
@@ -89,9 +89,9 @@ public class Indicator : View, Iindicator {
                 params?.addRule(RelativeLayout.CENTER_HORIZONTAL)
             } else {
                 params?.addRule(RelativeLayout.ALIGN_PARENT_RIGHT)
-                params?.rightMargin = ScreenUtil.dip2px(8f)
+                params?.rightMargin = ScreenUtil.dip2px(8f).toInt()
             }
-            params?.bottomMargin = ScreenUtil.dip2px(8f)
+            params?.bottomMargin = ScreenUtil.dip2px(8f).toInt()
         }
         return params!!
     }
